requirements.txt
speedtest-cli
pingparsing
colorama
argparse